moved sources and added test

+++ src/octopus/src/state/Handle.hh
@@ -6,7 +6,7 @@ namespace octopus
 
 /// @brief Handle is used to get an instance of the object in a State
 /// handles are stored in step to access the data to be modified
-typedef size_t Handle;
+typedef unsigned long Handle;
 
 }
